[Bronze V] Title: Multiply, Time: 0 ms, Memory: 2020 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main(void) {
-    int n, m, a, b;
+    long long n, m, a, b;
     cin >> n >> m;
     cin >> a;
     cin >> b;
